requirements.txt
fastapi
passlib[bcrypt]
python-jose[cryptography]